Rename the pkg
