+++ components/telemetry-operator/internal/configchecksum/hash.go
@@ -1,4 +1,4 @@
-package hash
+package configchecksum
 
 import (
 	"crypto/sha256"
