+++ components/telemetry-operator/internal/configchecksum/hash_test.go
@@ -1,4 +1,4 @@
-package hash
+package configchecksum
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
